requirements.txt
docopt
python-telegram-bot